Dashboard route, styles folder, renamed files starting with caps to lower case
