--- src/components/Login.css
@@ -1,26 +1,0 @@
-.login_container{
-    width: auto;
-    height: 100vh;
-    background: #ffffff;
-    display: flex;
-    justify-content: space-evenly;
-    padding: 0 10%;
-  
-}
-.login_left{
-    width: 100% ;
-    height: 100%;
-    background: #ffffff;
-    display: flex;
-    flex-direction: column;
-    background-color: palevioletred
-
-}
-.login_right{
-    width: 100% ;
-    height: 100%;
-    background: #ffffff;
-    display: flex;
-    flex-direction: column;
-    background-color: aqua;
-}
