+++ src/components/dashboard.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Dashboard() {
+    return (
+        <div>Dashboard</div>
+    )
+}
