+++ src/components/errorPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function ErrorPage() {
+    return (
+        <div>
+            <h1>oops!</h1>
+            <p>We could not find what you are looking for</p>
+        </div>
+    )
+}
