+++ src/styles/login.css
@@ -5,7 +5,6 @@
     display: flex;
     justify-content: space-evenly;
     padding: 0 10%;
-  
 }
 .login_left{
     width: 100% ;
