+++ src/styles/signup.css
@@ -1,6 +1,6 @@
 /* signup styles */
 .singup--page{
-    margin-top: 0;
+    margin: -8px ;
     background-image: url('../images/background.webp');
     background-size:cover;
     height: 100vh;
@@ -30,11 +30,15 @@
     background-color: #ffffff;
     color: black;
     width: 50%;
+    margin: auto;
     padding: 20px;
 }
 .right--top{
     text-align: center;
 
+}
+.right p{
+    font-size: 13px;
 }
 .singup--form{
     width: 80%;
@@ -69,15 +73,17 @@
     width: 40vw;
     justify-content: space-between;
     margin: auto;
+    margin-top: 4vh;
 }
 
 
 .a{
-    color: black;
     text-decoration: none;
-    font-size: 12px;
+    font-size: 16px;
     color: white;
-
+    font-weight: bold;
+    text-transform: capitalize;
+    /* border-bottom: 1px solid; */
 }
 
 .valid{
